handling promises

+++ src/app/products/shared/product.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import {AngularFirestore} from '@angular/fire/firestore';
-import {Observable} from 'rxjs';
+import {from, Observable, throwError} from 'rxjs';
 import {Product} from './product.model';
-import {map, tap} from 'rxjs/operators';
+import {first, map, switchMap, tap} from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
@@ -32,9 +32,41 @@ export class ProductService {
       );
   }
 
-  deleteProduct(id: string): Promise<void> {
+  deleteProduct(id: string): Observable<Product> {
     return this.db.doc<Product>('products/' + id)
-      .delete();
+      .get()
+      .pipe(
+        first(),
+        tap(productDocument => {
+          debugger;
+        }),
+        switchMap(productDocument => {
+          if (!productDocument || !productDocument.data()) {
+            throw new Error('Product not found');
+            debugger;
+          } else {
+            return from(
+              this.db.doc<Product>('products/' + id)
+                .delete()
+            ).pipe(
+              map(() => {
+                const data = productDocument.data() as Product;
+                data.id = productDocument.id;
+                return data;
+              })
+            );
+          }
+        })
+      );
+    /*return Observable.create(obs => {
+      this.db.doc<Product>('products/' + id)
+        .delete()
+        .then(() => obs.next())
+        .catch(err => obs.error(err))
+        .finally(() => obs.complete());
+    });*/
+    /*return this.db.doc<Product>('products/' + id)
+      .delete();*/
   }
 
   add10Products() {
